refactor: Abstract functions related to money
